Crate opg

Operator Precedence Grammar Parser

opg reads an context-free grammar input and outputs the precedence of the operators.

Modules

dfs
dfs

table
table

Structs

Production
A struct to represent a production.

Functions

find_eq
Find the equal operators

find_greater
Find the greater relations

find_less
Find the less relations

gen_firstvt
Generate FIRSTVT set for every non-terminals.

gen_lastvt
Generate LASTVT set for every non-terminals.

gen_productions
Generate production list for the grammar contents.

get_non_terminals
Get all the non terminals from the generated production.

get_terminals
Get terminals

main

opg_generate
Generate Operator Precedence Table for context-free grammar contents.