Function opg::get_non_terminals

pub(crate) fn get_non_terminals(
    productions: &Vec<Production>
) -> HashSet<String>

Get all the non terminals from the generated production.

Input

productions the vector of struct Production

Output

The hashset contains non-terminals.

Example

let nts = get_non_terminals(&productions);

Principles

Collect all the symbol on the left side in the productions.